fix: range selection in pool's getTransactions()
The logic of the previous code about returning transactions in the range
[start, start + size) was flawed - transactions would be removed from
the list without decrementing the counter i. For example, if:
* allTransactions is t1, t2, t3, t4, t5
* t1 and t2 are invalid (wouldn't pass validateTransactions() and it
  would remove them from the pool)
* start=1, size=2 is provided
then, t4 and t5 should be returned because after dropping the invalid
transactions the list becomes t3, t4, t5. But the code would wrongly
return t3 and t4.

Fix that by validating every transaction as it is encountered, skipping
invalid ones from the counting.